# Rewrite PK and modify Delta