--- a/graph/kv/kvtest/kvtest.go
+++ b/graph/kv/kvtest/kvtest.go
@@ -19,6 +19,7 @@ type Config struct{}
 
 func (c Config) quadStore() *graphtest.Config {
 	return &graphtest.Config{
+		NoPrimitives:            true,
 		SkipNodeDelAfterQuadDel: true,
 		SkipIntHorizon:          true,
 	}