--- a/graph/leveldb/migrate.go
+++ b/graph/leveldb/migrate.go
@@ -18,6 +18,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"strconv"
+	"time"
 
 	"github.com/cayleygraph/cayley/clog"
 	"github.com/cayleygraph/cayley/graph"
@@ -151,6 +152,7 @@ func upgrade1To2(db *leveldb.DB) error {
 	{
 		fmt.Println("Upgrading bucket d")
 		it := db.NewIterator(&util.Range{Start: []byte{'d'}, Limit: []byte{'d' + 1}}, nil)
+		h := int64(0)
 		for it.Next() {
 			k, v := it.Key(), it.Value()
 			id, err := strconv.ParseInt(string(k[1:]), 16, 64)
@@ -162,7 +164,8 @@ func upgrade1To2(db *leveldb.DB) error {
 			if err := json.Unmarshal(v, &val); err != nil {
 				return err
 			}
-			p := deltaToProto(val)
+			h++
+			p := deltaToProto(val, h, time.Now())
 			nv, err := p.Marshal()
 			if err != nil {
 				return err