--- a/graph/memstore/all_iterator.go
+++ b/graph/memstore/all_iterator.go
@@ -19,72 +19,144 @@ import (
 	"github.com/cayleygraph/cayley/graph/iterator"
 )
 
+var _ graph.Iterator = (*AllIterator)(nil)
+
 type AllIterator struct {
-	iterator.Int64
-	qs *QuadStore
+	uid  uint64
+	tags graph.Tagger
+
+	qs    *QuadStore
+	maxid int64 // id of last observed insert (prim id)
+	nodes bool
+
+	i    int // index into qs.all
+	cur  *primitive
+	done bool
 }
 
-type (
-	nodesAllIterator AllIterator
-	quadsAllIterator AllIterator
-)
-
-func newNodesAllIterator(qs *QuadStore) *nodesAllIterator {
-	var out nodesAllIterator
-	out.Int64 = *iterator.NewInt64(1, qs.nextID-1, true)
-	out.qs = qs
-	return &out
-}
-
-// No subiterators.
-func (it *nodesAllIterator) SubIterators() []graph.Iterator {
-	return nil
-}
-
-func (it *nodesAllIterator) Next() bool {
-	for {
-		if !it.Int64.Next() {
-			return false
-		}
-		_, ok := it.qs.revIDMap[int64(it.Int64.Result().(iterator.Int64Node))]
-		if !ok {
-			continue
-		}
-		return true
+func newAllIterator(qs *QuadStore, nodes bool, maxid int64) *AllIterator {
+	return &AllIterator{
+		uid: iterator.NextUID(),
+		qs:  qs, nodes: nodes,
+		i: -1, maxid: maxid,
 	}
 }
 
-func (it *nodesAllIterator) Err() error {
+func (it *AllIterator) Clone() graph.Iterator {
+	it2 := newAllIterator(it.qs, it.nodes, it.maxid)
+	it2.tags.CopyFrom(it)
+	return it2
+}
+
+func (it *AllIterator) Reset() {
+	it.i = -1
+	it.cur = nil
+	it.done = false
+}
+
+func (it *AllIterator) ok(p *primitive) bool {
+	if p.ID > it.maxid {
+		return false
+	} else if it.nodes && p.Value != nil {
+		return true
+	} else if !it.nodes && p.internalQuad != (internalQuad{}) {
+		return true
+	}
+	return false
+}
+
+func (it *AllIterator) Next() bool {
+	it.cur = nil
+	if it.done {
+		return false
+	}
+	all := it.qs.all
+	if it.i >= len(all) {
+		it.done = true
+		return false
+	}
+	it.i++
+	for ; it.i < len(all); it.i++ {
+		p := all[it.i]
+		if p.ID > it.maxid {
+			break
+		}
+		if it.ok(p) {
+			it.cur = p
+			return true
+		}
+	}
+	it.done = true
+	return false
+}
+
+func (it *AllIterator) Contains(v graph.Value) bool {
+	it.cur = nil
+	if it.done {
+		return false
+	}
+	id, ok := asID(v)
+	if !ok {
+		return false
+	}
+	p := it.qs.prim[id]
+	if p.ID > it.maxid {
+		return false
+	}
+	if !it.ok(p) {
+		return false
+	}
+	it.cur = p
+	return true
+}
+func (it *AllIterator) Result() graph.Value {
+	if it.cur == nil {
+		return nil
+	}
+	return graph.NewSequentialKey(it.cur.ID)
+}
+
+func (it *AllIterator) Err() error { return nil }
+func (it *AllIterator) Close() error {
+	it.done = true
 	return nil
 }
-
-func newQuadsAllIterator(qs *QuadStore) *quadsAllIterator {
-	var out quadsAllIterator
-	out.Int64 = *iterator.NewInt64(1, qs.nextQuadID-1, false)
-	out.qs = qs
-	return &out
+func (it *AllIterator) Tagger() *graph.Tagger {
+	return &it.tags
 }
 
-func (it *quadsAllIterator) Next() bool {
-	for {
-		if !it.Int64.Next() {
-			return false
-		}
-		i64 := int64(it.Int64.Result().(iterator.Int64Quad))
-		if int(i64) >= len(it.qs.log) {
-			return false
-		}
-		if it.qs.log[i64].DeletedBy != 0 || it.qs.log[i64].Action == graph.Delete {
-			continue
-		}
-		return true
+func (it *AllIterator) TagResults(dst map[string]graph.Value) {
+	for _, tag := range it.tags.Tags() {
+		dst[tag] = it.Result()
+	}
+
+	for tag, value := range it.tags.Fixed() {
+		dst[tag] = value
 	}
 }
 
-// Override Optimize from it.Int64 - it will hide our Next implementation in other cases.
+func (it *AllIterator) SubIterators() []graph.Iterator   { return nil }
+func (it *AllIterator) Optimize() (graph.Iterator, bool) { return it, false }
 
-func (it *nodesAllIterator) Optimize() (graph.Iterator, bool) { return it, false }
-func (it *quadsAllIterator) Optimize() (graph.Iterator, bool) { return it, false }
+func (it *AllIterator) UID() uint64 {
+	return it.uid
+}
+func (it *AllIterator) Type() graph.Type { return graph.All }
+func (it *AllIterator) Describe() graph.Description {
+	return graph.Description{
+		UID:  it.UID(),
+		Type: it.Type(),
+		Tags: it.tags.Tags(),
+	}
+}
+func (it *AllIterator) NextPath() bool { return false }
 
-var _ graph.Iterator = &nodesAllIterator{}
-var _ graph.Iterator = &quadsAllIterator{}
+func (it *AllIterator) Size() (int64, bool) {
+	// TODO: use maxid?
+	return int64(len(it.qs.all)), true
+}
+func (it *AllIterator) Stats() graph.IteratorStats {
+	st := graph.IteratorStats{NextCost: 1, ContainsCost: 1}
+	st.Size, st.ExactSize = it.Size()
+	return st
+}