Primary key was broken in many ways, for example it carried mutex, but all updates happened with Next that copied the value (with mutex).
Most backends are able to provide their own PK for log entries, so it's better to use it instead of emulating one.

Delta was changed in incompatible way: ID and Timestamp was removed.
As described above, ID should be set by QuadStore in case of successful after TX, while Timestamp should not be sent from client, since it's unreliable. Both values are now set in backend code.

Memstore was partially rewritten to reflect changes. It was also updated to use Primitives instead of nodes/quads.

Optimizations of memstore:
```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkNamePredicate-8                      197366        197153        -0.11%
BenchmarkLargeSetsNoIntersection-8            3241034       2018119       -37.73%
BenchmarkVeryLargeSetsSmallIntersection-8     5988325       4358409       -27.22%
BenchmarkHelplessContainsChecker-8            2116829       2061208       -2.63%
BenchmarkHelplessNotContainsFilms-8           695745        608946        -12.48%
BenchmarkHelplessNotContainsActors-8          6796812       5750716       -15.39%
BenchmarkNetAndSpeed-8                        807330        755597        -6.41%
BenchmarkKeanuAndNet-8                        627513        613111        -2.30%
BenchmarkKeanuAndSpeed-8                      711354        666505        -6.30%
BenchmarkKeanuOther-8                         2253536       1859483       -17.49%
BenchmarkKeanuBullockOther-8                  3951789       3321211       -15.96%
BenchmarkSaveBogartPerformances-8             308244        293941        -4.64%

benchmark                                     old allocs     new allocs     delta
BenchmarkNamePredicate-8                      458            452            -1.31%
BenchmarkLargeSetsNoIntersection-8            11652          6322           -45.74%
BenchmarkVeryLargeSetsSmallIntersection-8     26005          16231          -37.59%
BenchmarkHelplessContainsChecker-8            10868          8255           -24.04%
BenchmarkHelplessNotContainsFilms-8           3072           2475           -19.43%
BenchmarkHelplessNotContainsActors-8          40348          27421          -32.04%
BenchmarkNetAndSpeed-8                        4534           4144           -8.60%
BenchmarkKeanuAndNet-8                        3441           3323           -3.43%
BenchmarkKeanuAndSpeed-8                      3856           3561           -7.65%
BenchmarkKeanuOther-8                         11770          8539           -27.45%
BenchmarkKeanuBullockOther-8                  18591          13196          -29.02%
BenchmarkSaveBogartPerformances-8             1151           1079           -6.26%

benchmark                                     old bytes     new bytes     delta
BenchmarkNamePredicate-8                      28943         28174         -2.66%
BenchmarkLargeSetsNoIntersection-8            520249        400564        -23.01%
BenchmarkVeryLargeSetsSmallIntersection-8     1241740       950372        -23.46%
BenchmarkHelplessContainsChecker-8            758950        711689        -6.23%
BenchmarkHelplessNotContainsFilms-8           182826        166566        -8.89%
BenchmarkHelplessNotContainsActors-8          1313465       1158931       -11.77%
BenchmarkNetAndSpeed-8                        301581        286143        -5.12%
BenchmarkKeanuAndNet-8                        227941        220895        -3.09%
BenchmarkKeanuAndSpeed-8                      247413        236689        -4.33%
BenchmarkKeanuOther-8                         659190        584743        -11.29%
BenchmarkKeanuBullockOther-8                  966314        841981        -12.87%
BenchmarkSaveBogartPerformances-8             68837         65634         -4.65%
```